server/requirements.txt
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.1
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycparser==2.19
pymemcache==2.2.2
PyMySQL==0.9.3
pyserial==3.4
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
